Refactor layout and update SessionProvider import
